restructure fields
